Sanitized version of internal process doc
A copy of an internal doc I wrote in early 2022 to outline and standardize the usage of a set of scripts I wrote, to audit several discrete user lists against a master source.

Our IT department used this process to confirm that license counts were up-to-date on several subscription platforms. 

I wrote these scripts in powershell, performing source CSV manipulation and interfacing with several APIs and REST endpoints.